Backend/requirements.txt
flask
flask-socketio
flask-cors
python-dotenv
requests
gunicorn
eventlet